pyproject.toml
[project]
name = "it-copying"
version = "0.1.0"
description = "Evaluating generation of LMs pre/post intruction tuning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.34.2",
    "datasets>=3.0.1",
    "gin-config>=0.5.0",
    "ipykernel>=6.29.5",
    "torch>=2.4.1",
    "torchaudio>=2.4.1",
    "torchvision>=0.19.1",
    "transformers>=4.45.1",
]
